# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AI-Powered Smart Agriculture Advisory and Market Intelligence System for Ethiopia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weather/forecast:
    get:
      operationId: getWeatherForecast
      tags: [weather]
      summary: Get weather forecast and advisory
      parameters:
        - in: query
          name: region
          required: true
          schema:
            type: string
        - in: query
          name: days
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Weather forecast with advisory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherForecastResponse"
  /weather/advisory:
    get:
      operationId: getWeatherAdvisory
      tags: [weather]
      summary: Get smart weather-based farming advisory
      parameters:
        - in: query
          name: region
          required: true
          schema:
            type: string
        - in: query
          name: cropType
          schema:
            type: string
      responses:
        "200":
          description: Smart weather advisory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherAdvisoryResponse"
  /market/prices:
    get:
      operationId: getMarketPrices
      tags: [market]
      summary: Get current market prices
      parameters:
        - in: query
          name: commodity
          schema:
            type: string
        - in: query
          name: region
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
      responses:
        "200":
          description: Market price data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPriceList"
    post:
      operationId: submitMarketPrice
      tags: [market]
      summary: Submit market price report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketPriceSubmit"
      responses:
        "201":
          description: Price report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketPrice"
  /market/trends:
    get:
      operationId: getMarketTrends
      tags: [market]
      summary: Get market price trends and predictions
      parameters:
        - in: query
          name: commodity
          required: true
          schema:
            type: string
        - in: query
          name: period
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
      responses:
        "200":
          description: Market trends data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketTrendsResponse"
...
  /traceability/batches:
    get:
      operationId: listCropBatches
      tags: [traceability]
      summary: List crop batches with traceability data
      parameters:
        - in: query
          name: farmerId
          schema:
            type: integer
        - in: query
          name: cropType
          schema:
            type: string
      responses:
        "200":
          description: Crop batch list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropBatchList"
    post:
      operationId: createCropBatch
      tags: [traceability]
      summary: Create new crop batch record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCropBatchRequest"
      responses:
        "201":
          description: Batch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CropBatch"
  /traceability/batches/{id}/trace:
    get:
      operationId: traceCropBatch
      tags: [traceability]
      summary: Get full supply chain trace for a batch
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Supply chain trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupplyChainTrace"
...
components:
  schemas:
...
    WeatherForecast:
      type: object
      properties:
        date:
          type: string
        tempMin:
          type: number
        tempMax:
          type: number
        rainfall:
          type: number
        humidity:
          type: number
        windSpeed:
          type: number
        condition:
          type: string
        conditionAmharic:
          type: string
        conditionOromo:
          type: string
        farmingImpact:
          type: string
      required: [date, tempMin, tempMax, rainfall, humidity, windSpeed, condition, farmingImpact]
    WeatherForecastResponse:
      type: object
      properties:
        region:
          type: string
        current:
          $ref: "#/components/schemas/WeatherForecast"
        forecast:
          type: array
          items:
            $ref: "#/components/schemas/WeatherForecast"
        alerts:
          type: array
          items:
            type: string
        updatedAt:
          type: string
      required: [region, current, forecast, alerts, updatedAt]
    WeatherAdvisoryResponse:
      type: object
      properties:
        region:
          type: string
        cropType:
          type: string
        advisory:
          type: string
        advisoryAmharic:
          type: string
        advisoryOromo:
          type: string
        plantingRecommendation:
          type: string
        irrigationNeeded:
          type: boolean
        sprayingAdvisory:
          type: string
        harvestWindow:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
      required: [region, advisory, irrigationNeeded, riskLevel]
    MarketPrice:
      type: object
      properties:
        id:
          type: integer
        commodity:
          type: string
        commodityAmharic:
          type: string
        commodityOromo:
          type: string
        market:
          type: string
        region:
          type: string
        pricePerKg:
          type: number
        currency:
          type: string
          default: ETB
        unit:
          type: string
        change:
          type: number
        changePercent:
          type: number
        reportedAt:
          type: string
        source:
          type: string
      required: [id, commodity, market, region, pricePerKg, currency, unit, reportedAt, source]
    MarketPriceList:
      type: object
      properties:
        prices:
          type: array
          items:
            $ref: "#/components/schemas/MarketPrice"
        total:
          type: integer
        lastUpdated:
          type: string
      required: [prices, total, lastUpdated]
    MarketPriceSubmit:
      type: object
      properties:
        commodity:
          type: string
        market:
          type: string
        region:
          type: string
        pricePerKg:
          type: number
        unit:
          type: string
        reportedById:
          type: integer
      required: [commodity, market, region, pricePerKg, unit, reportedById]
    MarketTrendsResponse:
      type: object
      properties:
        commodity:
          type: string
        period:
          type: string
        dataPoints:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              avgPrice:
                type: number
              minPrice:
                type: number
              maxPrice:
                type: number
            required: [date, avgPrice, minPrice, maxPrice]
        prediction30Days:
          type: number
        trend:
          type: string
          enum: [rising, falling, stable]
        insights:
          type: array
          items:
            type: string
      required: [commodity, period, dataPoints, trend, insights]
...
    CropBatch:
      type: object
      properties:
        id:
          type: string
        farmerId:
          type: integer
        farmerName:
          type: string
        cropType:
          type: string
        variety:
          type: string
        quantityKg:
          type: number
        harvestDate:
          type: string
        region:
          type: string
        woreda:
          type: string
        grade:
          type: string
        blockchainHash:
          type: string
        blockchainTxId:
          type: string
        status:
          type: string
          enum: [harvested, processing, graded, sold, exported]
        createdAt:
          type: string
      required: [id, farmerId, farmerName, cropType, quantityKg, harvestDate, region, status, createdAt]
    CropBatchList:
      type: object
      properties:
        batches:
          type: array
          items:
            $ref: "#/components/schemas/CropBatch"
        total:
          type: integer
      required: [batches, total]
    CreateCropBatchRequest:
      type: object
      properties:
        farmerId:
          type: integer
        cropType:
          type: string
        variety:
          type: string
        quantityKg:
          type: number
        harvestDate:
          type: string
        region:
          type: string
        woreda:
          type: string
      required: [farmerId, cropType, quantityKg, harvestDate, region]
    SupplyChainEvent:
      type: object
      properties:
        eventType:
          type: string
        description:
          type: string
        location:
          type: string
        actor:
          type: string
        timestamp:
          type: string
        blockchainHash:
          type: string
        verified:
          type: boolean
      required: [eventType, description, location, actor, timestamp, verified]
    SupplyChainTrace:
      type: object
      properties:
        batchId:
          type: string
        cropType:
          type: string
        farmerName:
          type: string
        originRegion:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/SupplyChainEvent"
        currentStatus:
          type: string
        verificationUrl:
          type: string
      required: [batchId, cropType, farmerName, originRegion, events, currentStatus]
...